+++ src/assets/FieldAsset.php
@@ -255,6 +255,7 @@ class FieldAsset extends AssetBundle
                     'name' => Craft::t('neo', $blockType->name),
                     'handle' => $blockType->handle,
                     'maxBlocks' => $blockType->maxBlocks,
+                    'maxSiblingBlocks' => $blockType->maxSiblingBlocks,
                     'maxChildBlocks' => $blockType->maxChildBlocks,
                     'childBlocks' => is_string($blockType->childBlocks) ? Json::decodeIfJson($blockType->childBlocks) : $blockType->childBlocks,
                     'topLevel' => (bool)$blockType->topLevel,